creating window related hooks
