+++ hooks/useScreenWidth.tsx
@@ -1,0 +1,22 @@
+import { useEffect, useState } from "react";
+
+const useScreenWidth = () => {
+  const [screenWidth, setScreenWidth] = useState<number | undefined>();
+
+  useEffect(() => {
+    const handleResize = () => {
+      setScreenWidth(window.innerWidth);
+    };
+
+    handleResize();
+
+    window.addEventListener("resize", handleResize);
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  return screenWidth;
+};
+
+export default useScreenWidth;
