+++ hooks/useScrollY.tsx
@@ -1,0 +1,22 @@
+import { useEffect, useState } from "react";
+
+const useScrollY = () => {
+  const [scrollY, setScrollY] = useState(0);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setScrollY(window.scrollY);
+    };
+
+    handleScroll();
+
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
+
+  return scrollY;
+};
+
+export default useScrollY;
