+++ tsconfig.json
@@ -31,6 +31,7 @@
       "@/components/*": ["components/*"],
       "@/data/*": ["data/*"],
       "@/icons/*": ["../icons/*"],
+      "@/hooks/*": ["../hooks/*"],
       "@/types": ["types"],
       "@/constants": ["constants"]
     },
